test(complex-matcher): from Jest to test
